Remove unpublished rent-pending ATA builders
