+++ rust/pinocchio/src/instruction/mod.rs
@@ -1,15 +1,15 @@
 pub mod commit;
 pub mod commit_and_undelegate;
-pub mod create_rent_pending_ata;
 pub mod delegate;
 #[cfg(feature = "delegation-actions")]
 pub mod delegate_with_actions;
+pub mod rent_pending_ata;
 pub mod undelegate;
 
 pub use commit::*;
 pub use commit_and_undelegate::*;
-pub use create_rent_pending_ata::*;
 pub use delegate::*;
 #[cfg(feature = "delegation-actions")]
 pub use delegate_with_actions::*;
+pub use rent_pending_ata::*;
 pub use undelegate::*;
