+++ ts/kit/src/instructions/magic-program/index.ts
@@ -1,7 +1,7 @@
 export { createCommitInstruction } from "./scheduleCommit";
 export { createCommitAndUndelegateInstruction } from "./scheduleCommitAndUndelegate";
 export {
-  createRentPendingAtaInstruction,
   isRentPendingTokenAccount,
+  rentPendingAtaAddress,
   RENT_PENDING_ATA_CLOSE_AUTHORITY,
-} from "./createRentPendingAta";
+} from "./rentPendingAta";
