+++ ts/web3js/src/__test__/instructions.test.ts
@@ -14,8 +14,8 @@ import {
 import {
   createCommitInstruction,
   createCommitAndUndelegateInstruction,
-  createRentPendingAtaInstruction,
   isRentPendingTokenAccount,
+  rentPendingAtaAddress,
   RENT_PENDING_ATA_CLOSE_AUTHORITY,
 } from "../instructions/magic-program";
 import {
@@ -700,58 +700,17 @@ describe("Exposed Instructions (web3.js)", () => {
     });
   });
 
-  describe("createRentPendingAta instruction (Magic Program)", () => {
-    it("should create a CreateRentPendingAta instruction with required parameters", () => {
+  describe("rent-pending ATA helpers (Magic Program)", () => {
+    it("should derive the canonical ATA address", () => {
       const walletOwner = new PublicKey("11111111111111111111111111111113");
       const mint = new PublicKey("11111111111111111111111111111114");
       const [ata] = PublicKey.findProgramAddressSync(
-        [
-          walletOwner.toBuffer(),
-          TOKEN_PROGRAM_ID.toBuffer(),
-          mint.toBuffer(),
-        ],
+        [walletOwner.toBuffer(), TOKEN_PROGRAM_ID.toBuffer(), mint.toBuffer()],
         new PublicKey("ATokenGPvbdGVxr1b2hvZbsiqW5xWH25efTNsLJA8knL"),
       );
-      const instruction = createRentPendingAtaInstruction(
-        mockPublicKey,
-        walletOwner,
-        mint,
-      );
+      const derivedAta = rentPendingAtaAddress(walletOwner, mint);
 
-      expect(instruction.keys).toHaveLength(4);
-      expect(instruction.programId.toBase58()).toBe(
-        MAGIC_PROGRAM_ID.toBase58(),
-      );
-      expect(instruction.keys[0].pubkey.toBase58()).toBe(
-        mockPublicKey.toBase58(),
-      );
-      expect(instruction.keys[0].isSigner).toBe(true);
-      expect(instruction.keys[0].isWritable).toBe(true);
-      expect(instruction.keys[1].pubkey.toBase58()).toBe(ata.toBase58());
-      expect(instruction.keys[1].isWritable).toBe(true);
-      expect(instruction.keys[2].pubkey.toBase58()).toBe(mint.toBase58());
-      expect(instruction.keys[3].pubkey.toBase58()).toBe(
-        TOKEN_PROGRAM_ID.toBase58(),
-      );
-      expect(instruction.keys[3].isWritable).toBe(false);
-    });
-
-    it("should encode the discriminator and addresses", () => {
-      const walletOwner = new PublicKey("11111111111111111111111111111113");
-      const mint = new PublicKey("11111111111111111111111111111114");
-      const instruction = createRentPendingAtaInstruction(
-        mockPublicKey,
-        walletOwner,
-        mint,
-      );
-
-      expect(instruction.data.length).toBe(100);
-      expect(instruction.data.readUInt32LE(0)).toBe(15);
-      expect(instruction.data.subarray(4, 36)).toEqual(walletOwner.toBuffer());
-      expect(instruction.data.subarray(36, 68)).toEqual(mint.toBuffer());
-      expect(instruction.data.subarray(68, 100)).toEqual(
-        TOKEN_PROGRAM_ID.toBuffer(),
-      );
+      expect(derivedAta.toBase58()).toBe(ata.toBase58());
     });
 
     it("should detect rent-pending token accounts", () => {
